Legacy landing refactoring
- removed duplication from tests
- extracted functions
- created a struct to avoid passing so many parameters
