+++ handlers/legacy_landing_test.go
@@ -2,14 +2,17 @@ package handlers
 
 import (
 	"errors"
+	"github.com/ONSdigital/dp-api-clients-go/v2/files"
 	"github.com/ONSdigital/dp-api-clients-go/v2/zebedee"
 	"github.com/ONSdigital/dp-frontend-dataset-controller/config"
+	"github.com/ONSdigital/dp-frontend-dataset-controller/mapper"
 	coreModel "github.com/ONSdigital/dp-renderer/model"
 	"github.com/golang/mock/gomock"
-	"github.com/gorilla/mux"
 	. "github.com/smartystreets/goconvey/convey"
+	"io"
 	"net/http"
 	"net/http/httptest"
+	"strconv"
 	"testing"
 )
 
@@ -19,54 +22,39 @@ func TestLegacyLanding(t *testing.T) {
 	ctx := gomock.Any()
 	cfg := initialiseMockConfig()
 
+	mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
+	mockDatasetClient := NewMockDatasetClient(mockCtrl)
+	mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
+
 	Convey("test /data endpoint", t, func() {
 		Convey("test successful json response", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
 			mockZebedeeClient.EXPECT().Get(ctx, "12345", "/data?uri=/data").Return([]byte(`{"some_json":true}`), nil)
-			mockConfig := config.Config{}
 
 			w := httptest.NewRecorder()
-			req, err := http.NewRequest("GET", "/data", nil)
-			So(err, ShouldBeNil)
+			req, _ := http.NewRequest(http.MethodGet, "/data", nil)
 			req.AddCookie(&http.Cookie{Name: "access_token", Value: "12345"})
 
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil, mockConfig))
-
-			router.ServeHTTP(w, req)
+			LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil)(w, req)
 
 			So(w.Body.String(), ShouldEqual, `{"some_json":true}`)
 		})
 
-		Convey("test status 500 returned if zedbedee get returns error", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
+		Convey("test status 500 returned if zebedee get returns error", func() {
 			mockZebedeeClient.EXPECT().Get(ctx, userAuthToken, "/data?uri=/data").Return(nil, errors.New("something went wrong with zebedee"))
-			mockConfig := config.Config{}
 
 			w := httptest.NewRecorder()
-			req, err := http.NewRequest("GET", "/data", nil)
-			So(err, ShouldBeNil)
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil, mockConfig))
+			req, _ := http.NewRequest(http.MethodGet, "/data", nil)
 
-			router.ServeHTTP(w, req)
+			LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil)(w, req)
 
 			So(w.Code, ShouldEqual, http.StatusInternalServerError)
 		})
 	})
 
-	Convey("test legacylanding handler with non /data endpoint", t, func() {
+	Convey("test legacy landing handler with non /data endpoint", t, func() {
 		Convey("test successful data retrieval and rendering", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
-			mockConfig := config.Config{}
 			dlp := zebedee.DatasetLandingPage{URI: "https://helloworld.com"}
-			dlp.Datasets = append(dlp.Datasets, zebedee.Related{Title: "A dataset!", URI: "dataset.com"})
+			dlp.Datasets = append(dlp.Datasets, zebedee.Link{Title: "A dataset!", URI: "dataset.com"})
 
 			mockZebedeeClient.EXPECT().GetDatasetLandingPage(ctx, userAuthToken, collectionID, locale, "/somelegacypage").Return(dlp, nil)
 			mockZebedeeClient.EXPECT().GetBreadcrumb(ctx, userAuthToken, collectionID, locale, dlp.URI)
@@ -78,41 +66,26 @@ func TestLegacyLanding(t *testing.T) {
 			mockRend.EXPECT().BuildPage(gomock.Any(), gomock.Any(), "static")
 
 			w := httptest.NewRecorder()
-			req, err := http.NewRequest("GET", "/somelegacypage", nil)
-			So(err, ShouldBeNil)
+			req, _ := http.NewRequest(http.MethodGet, "/somelegacypage", nil)
 
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, mockRend, mockConfig))
-
-			router.ServeHTTP(w, req)
+			LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, mockRend).ServeHTTP(w, req)
 
 			So(w.Code, ShouldEqual, http.StatusOK)
 		})
 
 		Convey("test status 500 returned when zebedee client returns error retrieving landing page", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
-			mockConfig := config.Config{}
 			dlp := zebedee.DatasetLandingPage{}
 			mockZebedeeClient.EXPECT().GetDatasetLandingPage(ctx, userAuthToken, collectionID, locale, "/somelegacypage").Return(dlp, errors.New("something went wrong :("))
 
 			w := httptest.NewRecorder()
-			req, err := http.NewRequest("GET", "/somelegacypage", nil)
-			So(err, ShouldBeNil)
+			req, _ := http.NewRequest(http.MethodGet, "/somelegacypage", nil)
 
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil, mockConfig))
+			LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil).ServeHTTP(w, req)
 
-			router.ServeHTTP(w, req)
 			So(w.Code, ShouldEqual, http.StatusInternalServerError)
 		})
 
 		Convey("test status 500 returned when zebedee client returns error retrieving breadcrumb", func() {
-			mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
-			mockDatasetClient := NewMockDatasetClient(mockCtrl)
-			mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
-			mockConfig := config.Config{}
 			dlp := zebedee.DatasetLandingPage{URI: "https://helloworld.com"}
 			mockZebedeeClient.EXPECT().GetDatasetLandingPage(ctx, userAuthToken, collectionID, locale, "/somelegacypage").Return(dlp, nil)
 			mockZebedeeClient.EXPECT().GetBreadcrumb(ctx, userAuthToken, collectionID, locale, dlp.URI).Return(nil, errors.New("something went wrong"))
@@ -121,11 +94,124 @@ func TestLegacyLanding(t *testing.T) {
 			req, err := http.NewRequest("GET", "/somelegacypage", nil)
 			So(err, ShouldBeNil)
 
-			router := mux.NewRouter()
-			router.Path("/{uri:.*}").HandlerFunc(LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil, mockConfig))
+			LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, nil).ServeHTTP(w, req)
 
-			router.ServeHTTP(w, req)
 			So(w.Code, ShouldEqual, http.StatusInternalServerError)
+		})
+	})
+}
+
+func setupMockClients(ctx gomock.Matcher, mockZebedeeClient *MockZebedeeClient, mockRend *MockRenderClient, legacyURL string, dlp zebedee.DatasetLandingPage, authToken string, cfg config.Config) {
+	mockZebedeeClient.EXPECT().GetDatasetLandingPage(ctx, authToken, collectionID, locale, legacyURL).Return(dlp, nil)
+	mockZebedeeClient.EXPECT().GetBreadcrumb(ctx, authToken, collectionID, locale, dlp.URI)
+	mockZebedeeClient.EXPECT().GetHomepageContent(ctx, authToken, collectionID, locale, "/")
+
+	mockRend.EXPECT().NewBasePageModel().Return(coreModel.NewPage(cfg.PatternLibraryAssetsPath, cfg.SiteDomain))
+}
+
+func TestHandlersFilesAPI(t *testing.T) {
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+	ctx := gomock.Any()
+	cfg := initialiseMockConfig()
+
+	mockZebedeeClient := NewMockZebedeeClient(mockCtrl)
+	mockDatasetClient := NewMockDatasetClient(mockCtrl)
+	mockRend := NewMockRenderClient(mockCtrl)
+	mockFilesAPIClient := NewMockFilesAPIClient(mockCtrl)
+
+	landingPageURI := "https://helloworld.com"
+	dataSetURI := "dataset.com"
+	legacyURL := "/some_legacy_page"
+	expectedAuthToken := "auth-token"
+	authHeaderKey := "X-Florence-Token"
+
+	dlp := zebedee.DatasetLandingPage{
+		URI:      landingPageURI,
+		Datasets: []zebedee.Link{{Title: "Dataset", URI: dataSetURI}},
+	}
+
+	Convey("LegacyLanding handler file storage", t, func() {
+		Convey("File stored in Zebedee", func() {
+			expectedDownloadFileSize := "100"
+			expectedSupplementaryFileSize := "101"
+			downloadURI := "download_file_from_zebedee"
+			supplementaryURI := "supplementary_file_from_zebedee"
+
+			zebedeeDataset := zebedee.Dataset{
+				Downloads:          []zebedee.Download{{File: downloadURI, Size: expectedDownloadFileSize}},
+				SupplementaryFiles: []zebedee.SupplementaryFile{{File: supplementaryURI, Size: expectedSupplementaryFileSize}},
+			}
+
+			setupMockClients(ctx, mockZebedeeClient, mockRend, legacyURL, dlp, expectedAuthToken, cfg)
+
+			mockZebedeeClient.EXPECT().GetDataset(ctx, expectedAuthToken, collectionID, locale, dataSetURI).Return(zebedeeDataset, nil)
+
+			var actualPageModel mapper.StaticDatasetLandingPage
+
+			mockRend.EXPECT().BuildPage(gomock.Any(), gomock.Any(), "static").Do(func(w io.Writer, pageModel interface{}, templateName string) {
+				actualPageModel = pageModel.(mapper.StaticDatasetLandingPage)
+			})
+
+			w := httptest.NewRecorder()
+			req, _ := http.NewRequest("GET", legacyURL, nil)
+			req.Header.Set(authHeaderKey, expectedAuthToken)
+
+			handler := LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, mockRend)
+			handler(w, req)
+
+			actualDownloadSize := actualPageModel.DatasetLandingPage.Datasets[0].Downloads[0].Size
+			actualSupplementaryFileSize := actualPageModel.DatasetLandingPage.Datasets[0].SupplementaryFiles[0].Size
+
+			So(w.Code, ShouldEqual, http.StatusOK)
+			So(actualDownloadSize, ShouldEqual, expectedDownloadFileSize)
+			So(actualSupplementaryFileSize, ShouldEqual, expectedSupplementaryFileSize)
+		})
+
+		Convey("Files stored in Files API", func() {
+			expectedDownloadFileSize := "100"
+			expectedSupplementaryFileSize := "101"
+			downloadURI := "download_file_from_zebedee"
+			supplementaryURI := "supplementary_file_from_zebedee"
+			expectedDownloadFileSizeInt, _ := strconv.Atoi(expectedDownloadFileSize)
+			expectedSupplementaryFileSizeInt, _ := strconv.Atoi(expectedSupplementaryFileSize)
+			actualVersion := "v2"
+			zebedeeDownloadFileSize := ""
+			zebedeeSupplementaryFileSize := ""
+
+			fmdd := files.FileMetaData{SizeInBytes: uint64(expectedDownloadFileSizeInt)}
+			fmds := files.FileMetaData{SizeInBytes: uint64(expectedSupplementaryFileSizeInt)}
+
+			zebedeeDataset := zebedee.Dataset{
+				Downloads:          []zebedee.Download{{URI: downloadURI, Size: zebedeeDownloadFileSize, Version: actualVersion}},
+				SupplementaryFiles: []zebedee.SupplementaryFile{{URI: supplementaryURI, Size: zebedeeSupplementaryFileSize}},
+			}
+
+			setupMockClients(ctx, mockZebedeeClient, mockRend, legacyURL, dlp, expectedAuthToken, cfg)
+
+			mockZebedeeClient.EXPECT().GetDataset(ctx, expectedAuthToken, collectionID, locale, dataSetURI).Return(zebedeeDataset, nil)
+			mockFilesAPIClient.EXPECT().GetFile(ctx, gomock.Any(), expectedAuthToken).Return(fmdd, nil)
+			mockFilesAPIClient.EXPECT().GetFile(ctx, gomock.Any(), expectedAuthToken).Return(fmds, nil)
+
+			var actualPageModel mapper.StaticDatasetLandingPage
+
+			mockRend.EXPECT().BuildPage(gomock.Any(), gomock.Any(), "static").Do(func(w io.Writer, pageModel interface{}, templateName string) {
+				actualPageModel = pageModel.(mapper.StaticDatasetLandingPage)
+			})
+
+			w := httptest.NewRecorder()
+			req, _ := http.NewRequest("GET", legacyURL, nil)
+			req.Header.Set(authHeaderKey, expectedAuthToken)
+
+			handler := LegacyLanding(mockZebedeeClient, mockDatasetClient, mockFilesAPIClient, mockRend)
+			handler(w, req)
+
+			actualDownloadFileSize := actualPageModel.DatasetLandingPage.Datasets[0].Downloads[0].Size
+			actualSupplementaryFileSize := actualPageModel.DatasetLandingPage.Datasets[0].SupplementaryFiles[0].Size
+
+			So(w.Code, ShouldEqual, http.StatusOK)
+			So(actualDownloadFileSize, ShouldEqual, expectedDownloadFileSize)
+			So(actualSupplementaryFileSize, ShouldEqual, expectedSupplementaryFileSize)
 		})
 	})
 }
